requirements.txt
openai==0.28.0
streamlit>=1.28.0
python-dotenv>=1.0.0
requests>=2.31.0
markdown>=3.4.0
pygments>=2.15.0
